+++ test-e2e/cbor.test.ts
@@ -2,7 +2,6 @@ import { stop } from '@libp2p/interface'
 import * as cbor from 'cborg'
 import { createKuboRPCClient } from 'kubo-rpc-client'
 import { test, expect } from './fixtures/config-test-fixtures.ts'
-import { IPLD_CONVERSIONS } from './fixtures/ipld-conversions.ts'
 import { loadWithServiceWorker } from './fixtures/load-with-service-worker.ts'
 import type { KuboRPCClient } from 'kubo-rpc-client'
 import type { CID } from 'multiformats/cid'
@@ -52,20 +51,4 @@ test.describe('cbor', () => {
 
     expect(cbor.decode(await response?.body())).toStrictEqual(object)
   })
-
-  for (const conversion of IPLD_CONVERSIONS) {
-    // eslint-disable-next-line no-loop-func
-    test(`should return cbor block as ${conversion.format}`, async ({ page, baseURL }) => {
-      const response = await loadWithServiceWorker(page, `${baseURL}/ipfs/${cid}?format=${conversion.format}&download=true`)
-      expect(response.status()).toBe(200)
-
-      const headers = await response.allHeaders()
-      expect(headers['content-type']).toContain(conversion.mediaType)
-      expect(headers['cache-control']).toBe('public, max-age=29030400, immutable')
-      expect(headers['content-disposition']).toContain(conversion.disposition)
-      expect(headers['content-disposition']).toContain(conversion.filename(cid))
-
-      expect(await conversion.decode(response)).toStrictEqual(object)
-    })
-  }
 })